requirements.txt
numpy
opencv-python
Pillow
pillow-heif
open3d
matplotlib
scipy